style: clippy fixes + cargo fmt for pinned 1.94.0 toolchain
Auto-fixable clippy lints (to_vec, clone-on-Copy, collapsible-if), a clamp()
and an #[allow(too_many_arguments)] on an 11-param FFI export, plus a
workspace-wide cargo fmt to make fmt --check deterministic on the pin.

+++ crates/cpoe/src/store/mod.rs
@@ -455,7 +455,7 @@ pub fn check_disk_space(path: &Path) -> anyhow::Result<u64> {
                 std::io::Error::last_os_error()
             ));
         }
-        let available = stat.f_bavail as u64 * stat.f_frsize as u64;
+        let available = stat.f_bavail as u64 * stat.f_frsize;
         if available < MIN_FREE_SPACE_BYTES {
             anyhow::bail!(
                 "Insufficient disk space: {} MiB available, {} MiB required. \
